Fix: getDate 함수 유틸처리 불필요한 state삭제
