+++ src/pages/chatroom/ChatRoom.js
@@ -9,52 +9,19 @@ import {
   add_user_inputText_delete,
 } from '../../redux/action/inputChatAction';
 import { add_current_date } from '../../redux/action/dateNowAction';
+import { getClock } from '../../utils';
 
 export default function ChatRoom() {
   const dispatch = useDispatch();
   const scrollRef = useRef();
   const userdata = useSelector(state => state.login.userinfo);
-  const chatDatex = useSelector(state => state);
+  const chatDate = useSelector(state => state);
   const userChat = useSelector(state => state.input.chat);
   const [userMessage, setUserMessage] = useState('');
   const [ChatContents, setChatContents] = useState();
-  const [chatDate, setChatDate] = useState({
-    year: '',
-    month: '',
-    day: '',
-    hour: '',
-    minutes: '',
-    second: '',
-  });
-  const getClock = () => {
-    const date = new Date();
-    const year = date.getFullYear();
-    const month = String(date.getMonth() + 1).padStart(2, '0');
-    const day = String(date.getDate()).padStart(2, '0');
-    const hour = String(date.getHours()).padStart(2, '0');
-    const minutes = String(date.getMinutes()).padStart(2, '0');
-    const second = String(date.getSeconds()).padStart(2, '0');
-    setChatDate({
-      year: year,
-      month: month,
-      day: day,
-      hour: hour,
-      minutes: minutes,
-      second: second,
-    });
-    console.log(
-      `${chatDate.year}-${chatDate.month}-${chatDate.day} ${chatDate.hour}:${chatDate.minutes}:${chatDate.second}`
-    );
-    // console.log(
-    //   '현재시간',
-    //   `${year}-${month}-${day} ${hour}:${minutes}:${second}`
-    // );
-  };
-  useEffect(() => {
-    getClock();
-  }, [ChatContents]);
+  const date = getClock();
 
-  console.log('reducer', chatDatex);
+  console.log('reducer', chatDate);
   // console.log('userChat', userChat);
   // console.log('ChatContents', ChatContents);
   // console.log('userMessage', userMessage);
@@ -97,7 +64,7 @@ export default function ChatRoom() {
     if (e.key === 'Enter') {
       if (userMessage !== '') {
         dispatch(add_user_inputText(userInputData));
-        dispatch(add_current_date(chatDate));
+        dispatch(add_current_date(date));
         setUserMessage('');
       }
     }
